+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Note that the current nightly _requires_ VS Code Insiders.
 - feat: support setExpression for updating WATCH view variables ([#1075](https://github.com/microsoft/vscode-js-debug/issues/1075))
 - feat: integrate skipFiles with smartStepping to step through blackbox failures ([#1085](https://github.com/microsoft/vscode-js-debug/issues/1085))
 - fix: extension host not always being torn down when stopping debugging ([ref](https://github.com/microsoft/vscode/issues/126911))
+- fix: args list not updating when session is restarted ([ref](https://github.com/microsoft/vscode/issues/128058))
 
 ## v1.60 (August 2021)